Cleaned up compression algorythm.

+++ src/Avionics.h
@@ -63,7 +63,7 @@ private:
   void    printState();
   bool    logData();
   int16_t compressData();
-  int16_t compressVariable(float var, float minimum, float maximum, int16_t resolution, int32_t length);
+  int16_t compressVariable(float var, float minimum, float maximum, int16_t resolution, int16_t length);
 /*********************************  OBJECTS  **********************************/
   char COMMS_BUFFER[BUFFER_SIZE];
   DataFrame data;
